refactor: split military panel, players bar, production body, victory actions (#3878)
Continue colonizethis_app modularization by extracting library parts for
army expansion tiles, GP chip column/chip widgets, production screen body,
and victory overlay action row/laurel chrome without behavior changes.
