+++ app/lib/features/game/widgets/shell/game_map_players_bar.dart
@@ -10,6 +10,9 @@ import '../../../../config/editorial_monocle_palette.dart';
 import '../../screens/game/game_screen_shared.dart'
     show kGameMapPlayerChipKeyPrefix, kGameMapPlayersBarKey;
 
+part 'game_map_players_bar_chip.dart';
+part 'game_map_players_bar_chip_column.dart';
+
 /// Floating column of per–Great-Power chips on the in-game map stack.
 ///
 /// SPEC: `SPEC/ui/empire-overview.md` § Players bar. Issue #3898.
@@ -59,73 +62,6 @@ class GameMapPlayersBar extends StatelessWidget {
   static int powerScoreFor(Game game, String playerId) =>
       greatPowerPowerScore(game, playerId);
 
-  Color _swatchColorFor(
-    Map<String, (int r, int g, int b)> ownershipColors,
-    String playerId,
-  ) {
-    final tuple = ownershipColors[playerId];
-    if (tuple == null) {
-      return EditorialMonoclePalette.muted;
-    }
-    return Color.fromRGBO(tuple.$1, tuple.$2, tuple.$3, 1.0);
-  }
-
-  Widget _buildChipColumn(BuildContext context) {
-    final roster = greatPowerRoster(game);
-    if (roster.isEmpty) {
-      return const SizedBox.shrink();
-    }
-    final ownershipColors = factionOwnershipColorMapForOldWorld(game);
-    final scoreFormat = NumberFormat.decimalPattern('en_US');
-    final theme = Theme.of(context);
-    final mutedNameStyle =
-        (theme.textTheme.bodySmall ?? const TextStyle(fontSize: 12)).copyWith(
-          color: EditorialMonoclePalette.muted,
-          fontWeight: FontWeight.w500,
-          fontSize: 10,
-        );
-    final accentNameStyle = mutedNameStyle.copyWith(
-      color: EditorialMonoclePalette.accent,
-      fontWeight: FontWeight.w600,
-    );
-    final scoreStyle =
-        (theme.textTheme.bodySmall ?? const TextStyle(fontSize: 12)).copyWith(
-          color: EditorialMonoclePalette.accentDim,
-          fontFeatures: const <FontFeature>[FontFeature.tabularFigures()],
-          fontFamily: 'monospace',
-          fontSize: 9,
-        );
-    final chips = <Widget>[];
-    for (var i = 0; i < roster.length; i++) {
-      final player = roster[i];
-      if (i > 0) {
-        chips.add(const SizedBox(height: chipGap));
-      }
-      final highlight = highlightPlayerId != null &&
-          highlightPlayerId == player.id;
-      chips.add(
-        _PlayerChip(
-          key: Key('$kGameMapPlayerChipKeyPrefix${player.id}'),
-          name: player.displayName,
-          score: scoreFormat.format(powerScoreFor(game, player.id)),
-          swatchColor: _swatchColorFor(ownershipColors, player.id),
-          minWidth: chipMinWidth,
-          nameStyle: highlight ? accentNameStyle : mutedNameStyle,
-          scoreStyle: scoreStyle,
-        ),
-      );
-    }
-    return IgnorePointer(
-      ignoring: true,
-      child: Column(
-        key: kGameMapPlayersBarKey,
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.end,
-        children: chips,
-      ),
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
     final column = _buildChipColumn(context);
@@ -138,83 +74,6 @@ class GameMapPlayersBar extends StatelessWidget {
       top: top,
       right: right,
       child: column,
-    );
-  }
-}
-
-class _PlayerChip extends StatelessWidget {
-  const _PlayerChip({
-    super.key,
-    required this.name,
-    required this.score,
-    required this.swatchColor,
-    required this.minWidth,
-    required this.nameStyle,
-    required this.scoreStyle,
-  });
-
-  final String name;
-  final String score;
-  final Color swatchColor;
-  final double minWidth;
-  final TextStyle nameStyle;
-  final TextStyle scoreStyle;
-
-  @override
-  Widget build(BuildContext context) {
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topCenter,
-          end: Alignment.bottomCenter,
-          colors: <Color>[
-            EditorialMonoclePalette.surface,
-            EditorialMonoclePalette.bgDeep,
-          ],
-        ),
-        border: Border.all(
-          color: EditorialMonoclePalette.border,
-          width: 1,
-        ),
-      ),
-      child: ConstrainedBox(
-        constraints: BoxConstraints(minWidth: minWidth),
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 3),
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              Container(
-                width: 8,
-                height: 8,
-                decoration: BoxDecoration(
-                  color: swatchColor,
-                  border: Border.all(
-                    color: EditorialMonoclePalette.border,
-                    width: 1,
-                  ),
-                ),
-              ),
-              const SizedBox(width: 4),
-              Flexible(
-                child: Text(
-                  name,
-                  style: nameStyle,
-                  maxLines: 1,
-                  overflow: TextOverflow.ellipsis,
-                ),
-              ),
-              const SizedBox(width: 4),
-              Text(
-                score,
-                style: scoreStyle,
-                maxLines: 1,
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
