+++ app/lib/features/game/widgets/shell/game_map_players_bar_chip.dart
@@ -1,0 +1,78 @@
+part of 'game_map_players_bar.dart';
+
+class _PlayerChip extends StatelessWidget {
+  const _PlayerChip({
+    super.key,
+    required this.name,
+    required this.score,
+    required this.swatchColor,
+    required this.minWidth,
+    required this.nameStyle,
+    required this.scoreStyle,
+  });
+
+  final String name;
+  final String score;
+  final Color swatchColor;
+  final double minWidth;
+  final TextStyle nameStyle;
+  final TextStyle scoreStyle;
+
+  @override
+  Widget build(BuildContext context) {
+    return DecoratedBox(
+      decoration: BoxDecoration(
+        gradient: LinearGradient(
+          begin: Alignment.topCenter,
+          end: Alignment.bottomCenter,
+          colors: <Color>[
+            EditorialMonoclePalette.surface,
+            EditorialMonoclePalette.bgDeep,
+          ],
+        ),
+        border: Border.all(
+          color: EditorialMonoclePalette.border,
+          width: 1,
+        ),
+      ),
+      child: ConstrainedBox(
+        constraints: BoxConstraints(minWidth: minWidth),
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 3),
+          child: Row(
+            mainAxisSize: MainAxisSize.min,
+            mainAxisAlignment: MainAxisAlignment.end,
+            children: [
+              Container(
+                width: 8,
+                height: 8,
+                decoration: BoxDecoration(
+                  color: swatchColor,
+                  border: Border.all(
+                    color: EditorialMonoclePalette.border,
+                    width: 1,
+                  ),
+                ),
+              ),
+              const SizedBox(width: 4),
+              Flexible(
+                child: Text(
+                  name,
+                  style: nameStyle,
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
+                ),
+              ),
+              const SizedBox(width: 4),
+              Text(
+                score,
+                style: scoreStyle,
+                maxLines: 1,
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
